% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEcell.R
\name{MEcell}
\alias{MEcell}
\title{MEcell}
\usage{
MEcell(
  obj,
  assay = NULL,
  k_spatial = 16,
  k_nn = 20,
  usepca = F,
  K_adp = F,
  delta = 0.5,
  prune.SNN = 1/15
)
}
\arguments{
\item{obj}{A Seurat object}

\item{assay}{Assay to use in clustering. Default is 'NULL', that is the Default Assay of obj}

\item{k_spatial}{The number of spatial neighbors used to compute microenvironment profile. Default is 16}

\item{k_nn}{The number of nearest neighbors. Default is 20}

\item{usepca}{whether to use PCA scores to compute microenvironment profile. Default is FALSE. Set to True for whole transcriptome platform for speed.}

\item{K_adp}{whether to use adaptive strategy. Default is FALSE.}

\item{delta}{The parameter controlling the sensitivity to local microenvironment change. Default is 0.5,only valid when k_adp=T}

\item{prune.SNN}{The cutoff for acceptable Jaccard index when computing the neighborhood overlap for the SNN construction. Any edges with values less than or equal to this will be set to 0 and removed from the SNN graph. Default is 1/15}
}
\value{
A Seurat object with a microenvironment-refined SNN graph (named as "MEcell") stored in the respective slot.

library(MEcell)
obj<-readRDS(url("https://www.dropbox.com/s/f5khi1zperqkybg/Mouse_Brain_Serial_Section1_SagittalPosterior_rawimage_object.rds?dl=1"))
obj <- FindClusters(obj,graph.name="MEcell")
}
\description{
Leverage the microenvironment profile to improve spatial cell clustering
}
